requirements.txt
fastapi
uvicorn
python-jose
passlib
ldap3
jinja2
python-multipart
httpx
itsdangerous